fixed the bug of it not being the players turn
turns out there was for some reason a in isMoveable something that would automatically make it true if it was the first move.
it wasn't a problem at first because it was the first one to move (player 1). But after it was changed to player 2, it cause problems. So I took it out.

Don't know why it was in there in the first place.